+++ aslam_cv/aslam_time/include/aslam/implementation/Time.hpp
@@ -150,7 +150,7 @@ boost::posix_time::ptime TimeBase<T, D>::toBoost() const {
 #if defined(BOOST_DATE_TIME_HAS_NANOSECONDS)
   return pt::from_time_t(sec) + pt::nanoseconds(nsec);
 #else
-  return pt::from_time_t(sec) + pt::microseconds(nsec / 1000.0);
+  return pt::from_time_t(sec) + pt::microseconds(static_cast<int32_t>(nsec / 1000.0));
 #endif
 }
